requirements.txt
absl-py==0.7.1
adafruit-ampy==1.0.7
aiohttp==3.6.2
alabaster==0.7.12
alkana.py==0.0.1
altgraph==0.16.1
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
animatplot==0.4.1
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.1.2
async-timeout==3.0.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.5.1
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
beckett==0.8.0
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
boto3==1.9.134
botocore==1.12.134
Bottleneck==1.2.1
bottlenose==1.1.8
bz2file==0.98
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
colorama==0.4.1
conda==4.6.14
console==0.94
contextlib2==0.5.5
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
dask==1.1.4
datashape==0.5.4
decorator==4.4.0
defusedxml==0.5.0
discord.py==1.3.0
distributed==1.26.0
Django==2.2.2
docutils==0.14
ecdsa==0.15
emoji==0.5.4
entrypoints==0.3
esptool==2.8
et-xmlfile==1.0.1
ezenv==0.86
fastcache==1.0.2
fcache==0.4.7
Flask==1.0.2
Flask-Cors==3.0.7
future==0.16.0
gast==0.2.2
gensim==3.4.0
geoip2==2.9.0
get==1.0.3
gevent==1.4.0
gmpy2==2.0.8
google==2.0.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-images-download==2.8.0
google-pasta==0.1.6
googletrans==2.4.0
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
HTMLParser==0.0.2
httplib2==0.13.1
icrawler==0.6.2
idna==2.8
idna-ssl==1.1.0
imageio==2.5.0
imageio-ffmpeg==0.3.0
imagesize==1.1.0
importlib-metadata==0.0.0
inflect==0.2.5
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
iso8601==0.1.12
isort==4.3.16
itsdangerous==1.1.0
Janome==0.3.8
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
keyring==18.0.0
Kivy==1.10.1
Kivy-Garden==0.1.4
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
line-bot-sdk==1.5.0
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.2
macholib==1.11
Markdown==3.1.1
markovify==0.7.1
MarkupSafe==1.1.1
matplotlib==2.2.4
maxminddb==1.4.1
mccabe==0.6.1
mecab-python3==0.996.2
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==3.0.5
modulegraph==0.17
more-itertools==6.0.0
moviepy==1.0.0
mpmath==1.1.0
msgpack==0.6.1
multidict==4.7.4
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
neologdn==0.4
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
oathtool==1.0.0
oauth2client==4.1.3
oauthlib==3.0.1
odo==0.5.1
olefile==0.46
opencv-python==3.4.5.20
openpyxl==2.6.1
opt-einsum==3.1.0
oseti==0.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.10
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pefile==2018.8.8
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
PixivPy==3.3.6
pluggy==0.9.0
ply==3.11
pokepy==0.6.0
post==1.0.2
praw==6.3.1
prawcore==1.0.1
proglog==0.1.9
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.1
ptyprocess==0.6.0
public==1.0.3
py==1.8.0
py2app==0.17
pya3rt==1.1
pyaes==1.6.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
PyAudio==0.2.11
PyAutoGUI==0.9.38
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pydealer==1.4.0
PyDrive==1.3.1
pydub==0.23.1
pyfiglet==0.7.5
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.3.1
PyInstaller==3.4
pykakasi==0.96
pylint==2.3.1
pymlask==0.3.2
PyMsgBox==1.0.6
PyNaCl==1.3.0
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
PyPDF2==1.26.0
pyperclip==1.7.0
pyrsistent==0.14.11
PyScreeze==0.1.18
pyserial==3.4
PySocks==1.6.8
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python-dotenv==0.12.0
python-pptx==0.6.18
PyTweening==1.0.3
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
query-string==1.0.2
request==1.0.2
requests==2.22.0
requests-oauthlib==1.2.0
robobrowser==0.5.3
rope==0.12.0
rsa==4.0
ruamel-yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==0.19.1
seaborn==0.9.0
selenium==3.141.0
semidbm==0.5.1
Send2Trash==1.5.0
sengiri==0.2.1
sentencepiece==0.1.82
serial==0.0.69
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
slackbot==0.5.3
slacker==0.13.0
smart-open==1.8.2
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
SpeechRecognition==3.8.1
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
sqlparse==0.3.0
stanfordnlp==0.2.0
statsmodels==0.9.0
sympy==1.3
synthesizer==0.1.5
tables==3.5.1
tb-nightly==2.1.0a20191129
tblib==1.3.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-estimator-2.0-preview==2.0.0
tensorflow-hub==0.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
tf-nightly-2.0-preview==2.0.0.dev20191002
toolz==0.9.0
torch==1.1.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
tweepy==3.8.0
twitter==1.18.0
typed-ast==1.3.1
typing-extensions==3.7.2
unicodecsv==0.14.1
Unidecode==1.0.23
update-checker==0.16
uritemplate==3.0.0
urllib3==1.24.3
Wave==0.0.2
wcwidth==0.1.7
wdom==0.3.1
webencodings==0.5.1
websocket-client==0.56.0
websockets==6.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wikipedia==1.4.0
wrapt==1.11.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.4
xlwt==1.2.0
yarl==1.3.0
youtube-dl==2018.10.5
zict==0.1.4
zipp==0.3.3
